[Kernels] refactor QK kernels for reuse in M=1 kernels (#42178)
This is another step in the process to enable the M=1 kernels. The diff
looks horrible, but this is hoists the core tile logic out of
`_matmul_Q4_K` and `_matmul_Q6_K` to a standalone routine. This routine
then takes a lambda to compute the matmul using unpacked data. When the
M=1 kernel is activated, then there is a new code path that uses this
same standalone tile routine but using a lambda that uses packed data.

This also breaks out the common logic to accumulate and store the tile
output to the buffer.

QUANTIZATION_ORIG_REV_ID: 4bfd7a42b501f0272bab04a07ab43794155c1b13