[Kernels] Add GPU Fused layer norm operator (#42411)
Currently this PR adds 2 kernels
(a) One in which we can represent the rows in registers and a second
pass is not required for normalization
(b) A kernel which can apply to any sized input shapes but uses a second
pass to fetch again from global memory

Generally speaking the approach used in these Kernels is based on online
Welford algorithm which calculates the mean and variance in a single
pass + is numerically stable.

A followup PR(s)
-- would add in a third kernel in which Shared memory storage of input
data for reuse. This would sit in between the register and the global
memory approach.
-- hook the kernel with the MOGG layer so that the fused one gets picked
up and not the operation broken individually.
-- add benchmark comparisons so these kernels are tuned based on input
shape (and MOGG dispatches the right Kernel)

NN_ORIG_REV_ID: 927bd31a0e68452448e54ce57f1aa6209133856a