[MAX Driver] Pass runtime Device object into Model setup (#42473)
Towards GRA-594.

We want the graph compiler to use an externally managed Device object to
do allocations/copies/stream management. This allows the graph to be
executed on a Device created by the user. In this PR, we plumb the
Device pointer all the way down to Model::setup, but it does not get all
the way into the graph compiler yet, since a bit more work is required
on the graph compiler side to actually use this info.

Until GRA-629 is complete, or we add MAX Driver Python APIs, we don't
have a way to create the mojo "DeviceContext" (wrapper over CUDA driver)
from cpp (we can only create it from mojo), so the DeviceMap will
TEMPORARILY contain a nullptr when the DeviceRef key is "cuda". For now,
the graph compiler will need to default construct a DeviceContext when
it receives a nullptr in setup.

---------

Co-authored-by: Ian Tramble <ian@modular.com>
MOJO_MAX_ORIG_REV_ID: 0f2305ae8e884cb9665073e4d078eb588dd08897