Revert "[Serving] [SDK] Add initial mux inference test [SERV-152]" (#42960)
Reverts [Internal link]

This is triggering a breakage in the main. The CI runs on the PR are too
stale. We're fixing the underlying bug right now in Mojo async, but
reverting this PR until that fix lands to keep main green. This is
blocking presubmits

MOJO_MAX_ORIG_REV_ID: 476cf47cd131067a21831b7d1d8430b5eca90605