[graph-api] Add location information (#43045)
[graph-api] Add location information

This is a first attempt at reasonable location information.
It is not as good as a stack trace or pointer to specific line of code,
but it is reaonable.

The location information is 3 things:
1. The fully qualified layer name.
2. The op number within the parent layer.
3. The op number in this specific layer (does not include ops in sub
layers)

The location looks like this with mlir pretty printing on:
This is the first 3 ops of a GQA. The GQA layer is the 92nd op in the
model overall (it has no parent layer).
```mlir
    %182 = mo.constant {value = #M.dense_array<1, 0> : tensor<2xsi64>} : !mo.tensor<[2], si64> GQA:92:0
    %183 = rmo.mo.transpose(%88, %182) : (!mo.tensor<[5120, 3072], f32>, !mo.tensor<[2], si64>) -> !mo.tensor<[3072, 5120], f32> GQA:92:1
    %184 = rmo.matmul(%181, %183) : (!mo.tensor<[1, seq_len, 3072], f32>, !mo.tensor<[3072, 5120], f32>) -> !mo.tensor<[1, seq_len, 5120], f32> GQA:92:2
```

ref MSDK-531

MOJO_MAX_ORIG_REV_ID: 57abcebb45cdc07009599444a7e6932301b6358f