[Kernels] lift `num_kv_heads` and `head_size` to be parameters of KVCacheTypes (#43440)
The `num_heads` and `head_size` parameters need to be static for our
flash attention kernel to run the optimized implementation. We need this
to be a static dimension of the wrapped NDBuffer in the KVCache object,
which requires us to use this static shape information in the kernel.
This change adds the static parameter shapes to the types and kernel
interfaces. This is initially hard-coded to the tiny_stories llama2
model, but will be expanded to support other hyperparameters as a part
of RUNP-227

NN_ORIG_REV_ID: 5e0e2bf27b830a78363c80901322e08df8c4993c