[Kernels][Debugging] Add cuda-gdb tests (#43629)
This adds support for cuda-gdb tests that operate on mojo kernels tests
by adapting the infrastructure used for the existing mojo GPU tests.

With these changes, it's now possible to add llvm-lit-style check
directives to an existing gpu kernel test, and, if the file is symlinked
to `Kernels/test-gpu-debugging`, then cuda-gdb tests are run
automatically on it. This has the benefit of using the same test file
for a regular non-debug test and a debug test.

Symlinking is needed because a special test runner is used to simplify
writing of these tests, which forces us to have a completely new
llvm-lit configuration in a folder that is not
`Kernels/test`.

Besides allowing this new kind of test, which is shown for
`test_vec_add.mojo`, I'm moving the existing mojo GPU tests to this
folder because all GPU debugging tests need to run sequentially. This is
quite unfortunate but some hosts, like my mdcm machine, don't support
multiple cuda-gdb debug sessions. A solution I'm thinking for later is
to use a lock file-based mutex for these tests, so that they can be
defined anywhere and potentially allow for parallel compilation of the
underlying compiled mojo files.

GPU_TEST_ORIG_REV_ID: d9a4cbb8adc8d47496cb8296164692801b62f69f