[Kernels][GPU] Add load function that gives cache control (#44948)
This is similar to the load global operation in cutlass (albeit more
powerful). It allows one to control how the data gets loaded into cache.
For example, it would allow us to bypass the cache entirely if the data
is not reused (for example in A vector in a gemv).

Closes KERN-803

GPU_ORIG_REV_ID: 4c2721925b26080b0ca31cb55c03ec495ddf6046