[Kernels][GPU] Lift the tc reduction outside the loop (#45476)
This lifts the tc reduction outside the loop. To be able to do that
while still maintaining the numerical accuracy, we need to perform the
intermediate operations in s_type instead of a_type (this actually will
preserve more of accuracy). This does not really improve or degrade the
perf of the current implementation on the sizes we care about, but would
improve the numerics.

LINALG_ORIG_REV_ID: 4d90d1f940e9cc3edc7b73a06cdcedf9d736a65d