[MAX Driver] Add target info string to device descriptor (#45535)
Adds a target info string to the Device descriptor which contains the
compilation target triple, architecture, and a list of features
supported by the device.

For example, the descriptor for a CUDA device now looks like:
```
Device(type=cuda,gpu_id=0,target_info(triple=nvptx64-nvidia-cuda,arch=sm_89,features=[])
```

Updated tests to check for substrings instead of the whole descriptor
because the target_info depends on the device the target is running on,
and on CPU the target info is quite verbose.

---------

Co-authored-by: Ian Tramble <ian@modular.com>
MOJO_MAX_ORIG_REV_ID: 1085daae773595cb3056fa6979ec174566650889