[Kernels][GPU] Add cudnn bindings so that we can test against golden … (#43664)
…impl

This adds rough bindings to cudnn so that we are able to use in the
tests and benchmarks. This is useful for operations that do not exist in
cublass (e.g. norm). The entire thing is generated via

https://gist.github.com/abduld/dfaea97461a261ee084136b9609c6ccf

But there is a bunch of cleanup to do and we need to cleanup the enum
values as well. This is a good checkpoint however.

GPU_ORIG_REV_ID: 115d966d7710cd00a925eef8627a484353f53cfc