[Kernels][GPU] Remove dead code regarding setting constant memory (#45910)
The CUDA docs point you to the wrong path to setting constant memory by
stating that they are done via the module load functions. This is not
correct and are just done by looking up the symbol name in the module
and copying the data to it. As such, this code path is no longer valid
and we can just remove it.

GPU_ORIG_REV_ID: 3e1d9621e0c4c132375c392d6daba62c910cb876