[Kernels] Quick patch to avoid creating a DeviceContext on non-cuda Hosts (#45960)
This was causing issues while lifting Llama2 to use the `mo.opaque` kv
cache. The fix went in for matmul, but this PR was merged around the
same time and was missing the fix. This just patches that up.

NN_ORIG_REV_ID: 43fa7be082da7a5405f6cdf39902d2360c582725