[KGEN][MOGG] Fix Condition Annotating new DPS Interface (#46059)
One thing we need from MOGGAnnotate is the LIT name of argument type. We introspect these in later analysis (e.g. see an autoparameterized tensor? -- run autoparameterization).

The issue is the logic was a bit off in MOGGAnnotate. Here we simply forgot to annotate the `execute` and `shape` functions. We had an old block of code above the change which looks like it is supposed to do this, but as far as I am aware this only works on the old Tensor API interfaces. It happened to work before because the old block of code also annotates things with any tensor type.

With this, we properly annotate more functions. One final minor change is then needed -- certain arguments like variadics do not result in a proper annotation at this time -- instead returning a unit attribute. We make a change to handle this.

With this we can properly build the MOGG libraries without the pre-elaboration classes crashing when it includes future things that will be covered in "opaque type support" in the remainder of the stack.

Towards GRA-739

MOGG_ORIG_REV_ID: d4b7c898ebf260ec0892309d646158b15ff7fa15