[mojo-stdlib] Add a `as_noalias_ptr` method
The implementation of `exclusive=True` is not very robust. The metadata
can be dropped by the optimizer before it reaches LLVM. This introduces
a new `as_noalias_ptr` that is more robust and has better predictability
and composes better with other optimizations. The flag remains on the
type so that the behaviour of loads and stores can be altered.

LINALG_ORIG_REV_ID: 44b44dc0e958420c3da47408cb2b5a2ed260c1ab