[MOGG] Lower external constants to non tracked (#46191)
Lower external constants to a non-tracked buffer ref.
This currently works under the assumption that the caller of either the
model init or execute block is responsible for keeping the weights alive
for the duration of the call.
That responsibility is clarified in the engine API documentation.

Changes:
- Primitive to make an external call to create a TensorBufferRef with
  non-tracked memory.
- Case in lower-mogg-abi to create this op after
  mgp.buffer.constant.external.

Testing:
- A mo-opt test for the lower-mogg-abi change.
MOGG_ORIG_REV_ID: f5febd1c6ced390c37124dbd9790f8aa194b1d72