[GraphCompiler] Add support for propagation of exclusive flag for NDBuffer (#46312)
This change adds an `exclusive` flag to `NDBuffer` and updates the graph
compiler to propagate a `True` value to this flag to indicate that there
is no expected aliasing of this buffer.

A buffer is marked as exclusive when two conditions are met:
- The parent Kernel operation either does not have a `isViewOp`
attribute or has it set to false
- The defining op that created the `mo.tensor` instance is not a
`mogg.call` operation

These are the same conditions as those introduced in #45952 for stride
inference & propagation of `mo.tensor` instances.

Closes GRA-905

---------

Co-authored-by: Abdul Dakkak <adakkak@modular.com>
MOGG_ORIG_REV_ID: 623c64d8398669a8bdb564812dec8281c10012cb