[Kernels] add CPU version of `rms_norm` (#46558)
Add a basic kernel based on the CPU version of `layer_norm` and port the
GPU test to run on CPU.

None of the reduction generators quite did what was needed here to do a
sum of squares, so sorry for the manual loop. This can be revisited
later.

NN_ORIG_REV_ID: b0b9984fda03236c04c15813fcee71860de13803