[SDK] Give MojoValue a reference to the model (#46863)
Give the Python `MojoValue` a reference to the LoadedModel (<=>
ExecutableModel, which owns an MGP context).
Running the returned Mojo value’s destructor requires running a
jitted kernel owned by the MGP::Context in the ExecutableModel.
But the model was getting destroyed first in Python.
So as a heavyweight workaround for this awkwardness, give the MojoValue
a reference to the underlying ExecutableModel, to extend its lifetime
past the destruction of the MojoValue.

Fixes MSDK-912.

MAX_GRAPH_API_ORIG_REV_ID: 94ca8a0506ff571dacd517479c2545309f4e8460