[Kernels][GPU] More strict dispatch condition for GPU matmul. (#46936)
We're still assuming static K and N for dispatch multistage matmul but
only check B shape for it. This could break when some model doesn't fill
A and C with the static K, N values, e.g.

```
%637 = rmo.mo.matmul(%636, %67) : (!mo.tensor<[?, ?], f32>, !mo.tensor<[768, 768], f32>) -> !mo.tensor<[?, 768], f32>
```
This PR further constrain the dispatch on A with static K and C with
static N.

We can support fully dynamic shapes with runtime layout but it's not
urgent for now, so address it later.

LINALG_ORIG_REV_ID: 8ffa81a386094886473d5353ad19a927c808b733